Separate the open(url) and parse of html doc steps.
Separate the two function calls so that the metadata of the HTTP response
can be accessed. This is necessary because we need to use cookies sent
to us so that subsequent requests for pages are honoured (if no cookies
are sent in subsequent requests, we only get the first page every time).